Add configuration using redis.properties
    Turn on blocking io as connections would throw an error when pool was exhausted
    Shade configuration updated
