+++ pom.xml
@@ -34,6 +34,11 @@
 			<version>2.4.2</version>
 		</dependency>
 		<dependency>
+		  <groupId>commons-configuration</groupId>
+		  <artifactId>commons-configuration</artifactId>
+		  <version>1.10</version>
+		</dependency>
+		<dependency>
 			<groupId>redis.clients</groupId>
 			<artifactId>jedis</artifactId>
 			<version>2.8.0</version>
@@ -95,8 +100,8 @@
 							<shadedPattern>com.webonise.internal.org.apache.commons</shadedPattern>
 						</relocation>
 						<relocation>
-							<pattern>redis</pattern>
-							<shadedPattern>com.webonise.internal.redis</shadedPattern>
+							<pattern>redis.clients</pattern>
+							<shadedPattern>com.webonise.internal.redis.clients</shadedPattern>
 						</relocation>
 					</relocations>
 					<!-- put your configurations here -->
